pyproject.toml
[project]
name = "semantic-linker"
version = "0.1.0"
description = "Semantic record linkage and deduplication using embeddings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.0",
    "gemini-batch",
    "google-genai>=0.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[tool.uv.sources]
gemini-batch = { git = "https://github.com/phenschke/gemini-batch" }

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require GOOGLE_API_KEY)",
]
